requirements.txt
openai==1.3.7
google-generativeai==0.3.2
SpeechRecognition==3.10.0
gtts==2.4.0
pyttsx3==2.90
sounddevice==0.4.6
pyaudio==0.2.11
librosa==0.10.1
webrtcvad==2.0.10
streamlit==1.28.1
gradio==4.7.1
langchain==0.0.350
reportlab==4.0.7
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
python-dotenv==1.0.0
plotly==5.17.0
